+++ R/compartment_bioavailability.R
@@ -38,11 +38,11 @@ setMethod("get_name", signature = c("compartment_bioavailability"), definition =
 })
 
 #_______________________________________________________________________________
-#----                             getPrefix                                ----
+#----                             get_prefix                                ----
 #_______________________________________________________________________________
 
-#' @rdname getPrefix
-setMethod("getPrefix", signature = c("compartment_bioavailability"), definition = function(object, ...) {
+#' @rdname get_prefix
+setMethod("get_prefix", signature = c("compartment_bioavailability"), definition = function(object, ...) {
   dest <- process_extra_arg(args=list(...), name="dest", default="rxode2")
   if (dest=="mrgsolve") {
     return("F")
@@ -52,10 +52,10 @@ setMethod("getPrefix", signature = c("compartment_bioavailability"), definition 
 })
 
 #_______________________________________________________________________________
-#----                           getRecordName                               ----
+#----                           get_record_name                               ----
 #_______________________________________________________________________________
 
-#' @rdname getRecordName
-setMethod("getRecordName", signature = c("compartment_bioavailability"), definition = function(object) {
+#' @rdname get_record_name
+setMethod("get_record_name", signature = c("compartment_bioavailability"), definition = function(object) {
   return("F")
 })