+++ R/compartment_initial_condition.R
@@ -37,20 +37,20 @@ setMethod("get_name", signature=c("compartment_initial_condition"), definition=f
 })
 
 #_______________________________________________________________________________
-#----                             getPrefix                                ----
+#----                             get_prefix                                ----
 #_______________________________________________________________________________
 
-#' @rdname getPrefix
-setMethod("getPrefix", signature = c("compartment_initial_condition"), definition = function(object, ...) {
+#' @rdname get_prefix
+setMethod("get_prefix", signature = c("compartment_initial_condition"), definition = function(object, ...) {
   return("")
 })
 
 #_______________________________________________________________________________
-#----                           getRecordName                               ----
+#----                           get_record_name                               ----
 #_______________________________________________________________________________
 
-#' @rdname getRecordName
-setMethod("getRecordName", signature = c("compartment_initial_condition"), definition = function(object) {
+#' @rdname get_record_name
+setMethod("get_record_name", signature = c("compartment_initial_condition"), definition = function(object) {
   return("INIT")
 })