+++ R/global.R
@@ -6,12 +6,12 @@
 #' @return no return value
 #' @importFrom methods callNextMethod is new show slot slot<- slotNames validObject
 #' @keywords internal
-methodsGlobalFunctions <- function() {
+methods_global_functions <- function() {
   
 }
 
 #' @importFrom rlang .data :=
 #' @keywords internal
-rlangGlobal <- function() {
+rlang_global <- function() {
   
 }